+++ extensions/amp-story-auto-ads/0.1/test/test-story-ad-localization.js
@@ -1,0 +1,37 @@
+/**
+ * Copyright 2019 The AMP HTML Authors. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS-IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import {CtaTypes, StoryAdLocalization} from '../story-ad-localization';
+
+describes.realWin('amp-story-auto-ads:localization', {amp: true}, env => {
+  describe('story ad localization', () => {
+    it('registers and returns localized string [es]', () => {
+      const {win} = env;
+      win.document.documentElement.setAttribute('lang', 'es');
+      const localization = new StoryAdLocalization(env.win);
+      const localized = localization.getLocalizedString(CtaTypes['INSTALL']);
+      expect(localized).to.equal('Instalar ahora');
+    });
+
+    it('registers and returns localized string [zh-CN]', () => {
+      const {win} = env;
+      win.document.documentElement.setAttribute('lang', 'zh-CN');
+      const localization = new StoryAdLocalization(env.win);
+      const localized = localization.getLocalizedString(CtaTypes['SHOP']);
+      expect(localized).to.equal('选购');
+    });
+  });
+});
